+++ Code/AT_2Panels.lua
@@ -21,7 +21,8 @@ local iconATSection     = ModDir.."UI/Icons/ATSection.png"
 --  setup or tear down all the AT variables in a rocket
 function ATsetupVariables(rocket, init)
 	if init then
-		rocket.AT_enabled              = true  -- var used to turn system on/off
+		g_AT_NumOfTouristRockets = g_AT_NumOfTouristRockets + 1 -- increment the global counter
+		rocket.AT_enabled              = true   -- var used to turn system on/off
 		rocket.AT_departures           = rocket.AT_departures or 0      -- number of tourists returning to earth, keep departures if cycling button on/off
 		rocket.AT_arriving_tourists    = 0      -- number of tourists picked up from earth
 		rocket.AT_departuretime        = 0      -- gametime var for departure time
@@ -36,6 +37,7 @@ function ATsetupVariables(rocket, init)
 		rocket.AT_depart_thread        = false  -- var holds countdown thread for departures
 		rocket.AT_status_thread        = false  -- var holds status thread if it exists for boarding complete
 	else
+		g_AT_NumOfTouristRockets = g_AT_NumOfTouristRockets - 1
 	  rocket.AT_enabled              = nil
 		--rocket.AT_departures           = nil -- remmed, keep departures if cycling on/off
 		rocket.AT_arriving_tourists    = nil