+++ metadata.lua
@@ -1,8 +1,8 @@
 return PlaceObj('ModDef', {
 	'title', "Automated Tourism",
-	'description', "Automated Tourism\nv1.2.0\nby SKiRich",
+	'description', "Automated Tourism\nv1.3.0\nby SKiRich",
 	'image', "AutomatedTourism.png",
-	'last_changes', "v1.2.0 May 29th, 2019",
+	'last_changes', "v1.3.0 May 29th, 2019",
 	'ignore_files', {
 		"*.git/*",
 		"*.svn/*",
@@ -14,14 +14,14 @@ return PlaceObj('ModDef', {
 	'pops_desktop_uuid', "c515e2e3-ff0a-4dc2-91f4-c6b58431e755",
 	'pops_any_uuid', "44a35da1-f32e-4034-a8e5-4280e705a553",
 	'author', "SkiRich",
-	'version', 22,
+	'version', 25,
 	'lua_revision', 233360,
-	'saved_with_revision', 244677,
+	'saved_with_revision', 245618,
 	'code', {
 		"Code/AT_1Init.lua",
 		"Code/AT_2Panels.lua",
 		"Code/AT_3ModConfig.lua",
 	},
-	'saved', 1559097534,
+	'saved', 1559186692,
 	'screenshot1', "Images/AutomatedTourism.png",
 })